[WIP] Run against the official test suite
This imports tests from https://github.com/apache/orc/tree/main/examples
and adds tests which read using this crate, output JSON, and compare
with the expected results.

As a side-effect, I also amended write.py to write in that format, and
run the same kind of test on those.